refactor: move DatabaseException to its own file
